Optionally specify separate client and server breakpoint lists
Previously there was a single breakpoint list shared by client and server.
This change allows the client to specify additional breakpoints that are not cached on the server, but still bucket
client requests into common widths, reducing load on the server (since these images will be cached outside the server,
including CDN).
Ultimately this allows client to add breakpoints for larger images, with incurring disk storage cost for caching them
directly on server.